+++ tests/interface/TWEthereumAbiValueEncoderTests.cpp
@@ -4,7 +4,7 @@
 // terms governing use, modification, and redistribution, is contained in the
 // file LICENSE at the root of the source code distribution tree.
 
-#include <TrustWalletCore/TWEthereumAbiValueEncode.h>
+#include <TrustWalletCore/TWEthereumAbiValueEncoder.h>
 
 #include "Data.h"
 #include "HexCoding.h"
